+++ Clases/URLProcessor.vb
@@ -20,7 +20,9 @@ Public Class URLProcessor
             If URLExtractor.IsMegaFolder(URL) Then
                 Dim FolderID As String = URLExtractor.ExtraerFileID(URL)
                 Dim FolderKey As String = URLExtractor.ExtraerFileKey(URL)
-                For Each FileURL In MegaFolderHelper.RetrieveLinksFromFolder(FolderID, FolderKey)
+                Dim SubFolderID As String = URLExtractor.ExtraerSubFolderID(URL)
+                Dim SubFileID As String = URLExtractor.ExtraerSubFileID(URL)
+                For Each FileURL In MegaFolderHelper.RetrieveLinksFromFolder(FolderID, FolderKey, SubFolderID, SubFileID)
                     URLs3.Add(FileURL)
                 Next
             ElseIf URLExtractor.IsELC(URL) Then
@@ -29,7 +31,9 @@ Public Class URLProcessor
                     If URLExtractor.IsMegaFolder(FileURL) Then
                         Dim FolderID As String = URLExtractor.ExtraerFileID(FileURL)
                         Dim FolderKey As String = URLExtractor.ExtraerFileKey(FileURL)
-                        For Each FileURL2 In MegaFolderHelper.RetrieveLinksFromFolder(FolderID, FolderKey)
+                        Dim SubFolderID As String = URLExtractor.ExtraerSubFolderID(FileURL)
+                        Dim SubFileID As String = URLExtractor.ExtraerSubFileID(FileURL)
+                        For Each FileURL2 In MegaFolderHelper.RetrieveLinksFromFolder(FolderID, FolderKey, SubFolderID, SubFileID)
                             URLs3.Add(New FileURL(Fichero.HIDDEN_LINK & FileURL2.URL, FileURL2.Path))
                         Next
                     Else